# specs/001-web-auth-todo/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: API for managing todos with authentication
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

paths:
  /api/todos:
    get:
      summary: Get user's todos
      description: Retrieve all todos belonging to the authenticated user
      responses:
        '200':
          description: Successfully retrieved user's todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized - invalid or missing JWT token
      security:
        - bearerAuth: []

    post:
      summary: Create a new todo
      description: Create a new todo for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Bad request - invalid input
        '401':
          description: Unauthorized - invalid or missing JWT token
      security:
        - bearerAuth: []

  /api/todos/{id}:
    patch:
      summary: Update a todo
      description: Update an existing todo for the authenticated user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The todo ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Bad request - invalid input
        '401':
          description: Unauthorized - invalid or missing JWT token
        '404':
          description: Todo not found
      security:
        - bearerAuth: []

    delete:
      summary: Delete a todo
      description: Delete an existing todo for the authenticated user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The todo ID
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          description: Unauthorized - invalid or missing JWT token
        '404':
          description: Todo not found
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - priority
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the todo
        title:
          type: string
          maxLength: 200
          description: Title of the todo
        description:
          type: string
          maxLength: 1000
          description: Optional description of the todo
        priority:
          type: string
          enum: [high, medium, low]
          description: Priority level of the todo
        completed:
          type: boolean
          description: Whether the todo is completed
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns this todo
        created_at:
          type: string
          format: date-time
          description: Timestamp when the todo was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the todo was last updated

    CreateTodoRequest:
      type: object
      required:
        - title
        - priority
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        priority:
          type: string
          enum: [high, medium, low]

    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        priority:
          type: string
          enum: [high, medium, low]
        completed:
          type: boolean